+++ CHANGELOG.md
@@ -154,6 +154,7 @@
 - PR #4407 Enable `.str.slice` & `.str.get` and `.str.zfill` unit-tests
 - PR #4412 Require Dask + Distributed 2.12.0+
 - PR #4377 Support loading avro files that contain nested arrays
+- PR #4405 Port nvstrings (Sub)string Comparisons functions to cuDF Python/Cython
 
 ## Bug Fixes
 
@@ -236,6 +237,7 @@
 - PR #4406 Fix sorted merge issue with null values and ascending=False
 - PR #4423 Tighten up Dask serialization checks
 - PR #4438 Fix repl-template error for replace_with_backrefs
+- PR #4434 Fix join_strings logic with all-null strings and non-null narep
 
 
 # cuDF 0.12.0 (04 Feb 2020)